% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barplot_selectionProbability}
\alias{barplot_selectionProbability}
\title{Barplot Selection Probabilities}
\usage{
barplot_selectionProbability(stabs_object, ylim = c(0, 1),
  onlySelected = TRUE, las = 2, ...)
}
\arguments{
\item{stabs_object}{the \code{stabs} object resulting from stability selection.}

\item{ylim}{the limits for the y-axis.}

\item{onlySelected}{logical (default=TRUE) indicating if only selected predictors' selection probabilities 
should be plotted.}

\item{las}{(2 by default) plot labels vertically or horizontally.}

\item{...}{additional parameters for the \code{barplot} function.}
}
\value{
barplot of selection probabilities.
}
\description{
Do a bar plot of the selection probabilities in order from greatest to least.
}
