% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{buildDirGraphFromKmers}
\alias{buildDirGraphFromKmers}
\title{Build a directed graph from enriched k-mers}
\usage{
buildDirGraphFromKmers(seqs, x, BPPARAM = SerialParam(), includeRevComp = TRUE)
}
\arguments{
\item{seqs}{A \code{\link{DNAStringSet}} with sequences within which the 
kmers in \code{x} are enriched.}

\item{x}{A \code{character} vector of k-mers enriched in the sequences 
in \code{seqs}.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}

\item{includeRevComp}{A \code{logical} scalar. If \code{TRUE} (default),
concatenate \code{seqs} and their reverse-complemented versions
before building the graph. The choice here should correspond to the choice
made when extracting the k-mers.}
}
\value{
A directed graph with nodes corresponding to k-mers.
}
\description{
Build a directed graph from k-mers enriched in a set of sequences. The 
graph is based on a de Bruijn graph, in which edges are weighted based on 
the co-occurrence of neighboring k-mers in the provided input sequences. 
Putative motifs can be found by only retaining edges with weight above a 
user-defined threshold (see \code{getMotifsFromDirGraph}).
}
\examples{
\dontrun{
g <- buildDirGraphFromKmers(seqs, x)
hist(E(g)$weights)
getMotifsFromDirGraph(g, 10)
}

}
