% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_wrapper.R
\name{calcBinnedMotifEnr}
\alias{calcBinnedMotifEnr}
\title{Binned motif enrichment analysis.}
\usage{
calcBinnedMotifEnr(
  seqs,
  bins,
  motifs,
  method = c("R", "Homer"),
  pseudocount.log2enr = 8,
  pseudocount.pearsonResid = 0.001,
  p.adjust.method = "BH",
  BPPARAM = SerialParam(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{seqs}{The sequences to analyze. Either a \code{DNAStringSet} (for
\code{method = "R"}) or a \code{GRanges} object defining the genomic
regions from which to extract the sequences (for \code{method = "Homer"}).}

\item{bins}{Factor of the same length and order as \code{seqs}, indicating
the bin for each sequence or genomic region. Typically the return value of
\code{\link[monaLisa]{bin}}.}

\item{motifs}{The motifs for which to calculate enrichments. Either a
\code{PWMatrixList} (for \code{method = "R"}) or a file with motifs in
Homer format (for \code{method = "Homer"}).}

\item{method}{A \code{character} scalar specifying the backend to use for
enrichment calculations. One of \code{"R"} (default) or \code{"Homer"}.}

\item{pseudocount.log2enr}{A numerical scalar with the pseudocount to add to
foreground and background counts when calculating log2 motif enrichments}

\item{pseudocount.pearsonResid}{A numerical scalar with the pseudocount to add
to foreground and background frequencies when calculating expected counts
and Pearson residuals.}

\item{p.adjust.method}{A character scalar selecting the p value adjustment
method (used in \code{\link[stats]{p.adjust}}).}

\item{BPPARAM}{Specifies the number of CPU cores to use for parallel
processing. Either a \code{\link[BiocParallel]{SerialParam}} or a
\code{\link[BiocParallel]{MulticoreParam}} object instance, or for
\code{method = "Homer"} a numeric scalar.}

\item{verbose}{A logical scalar. If \code{TRUE}, print progress messages.}

\item{...}{Additional arguments for passed to
\code{\link{calcBinnedMotifEnrR}} or
\code{\link{calcBinnedMotifEnrHomer}}.}
}
\value{
A \code{SummarizedExperiment} object with motifs in rows and bins
  in columns, containing six assays: \itemize{
  \item{negLog10P}{: -log10 P values}
  \item{negLog10Padj}{: -log10 adjusted P values}
  \item{pearsonResid}{: motif enrichments as Pearson residuals}
  \item{log2enr}{: motif enrichments as log2 ratios}
  \item{sumForegroundWgtWithHits}{: Sum of foreground sequence weights
    in a bin that have motif hits}
  \item{sumBackgroundWgtWithHits}{: Sum of background sequence weights
    in a bin that have motif hits}
}
}
\description{
Calculate motif enrichment analysis on bins of sequences.
  Motif occurrences in each bin are contrasted to the motif occurrences
  in a background set of sequences (by default the sequences in all other
  bins).
}
\details{
This function is a wrapper for
  \code{\link{calcBinnedMotifEnrR}} or
  \code{\link{calcBinnedMotifEnrHomer}}. For additional supported arguments
  (via \code{...}) please see the help pages of these functions.
}
\examples{
seqs <- DNAStringSet(c("GTCAGTCGATC", "CAGTCTAGCTG",
                       "CGATCGTCAGT", "AGCTGCAGTCT"))
bins <- factor(rep(1:2, each = 2))
m <- rbind(A = c(2, 0, 0),
           C = c(1, 1, 0),
           G = c(0, 2, 0),
           T = c(0, 0, 3))
pwms <- PWMatrixList(PWMatrix(ID = "m1", profileMatrix = m),
                     PWMatrix(ID = "m2", profileMatrix = m[, 3:1]))
calcBinnedMotifEnr(seqs = seqs, bins = bins, motifs = pwms,
                   method = "R", min.score = 3)

}
\seealso{
\code{\link{bin}} for binning of sequences;
  \code{\link{calcBinnedMotifEnrR}} that is used for
  \code{method = "R"} and \code{\link{calcBinnedMotifEnrHomer}} used
  for \code{method = "Homer"}.
}
