% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{calculate_GC_weight}
\alias{calculate_GC_weight}
\title{Get GC weights for background regions}
\usage{
calculate_GC_weight(df)
}
\arguments{
\item{df}{a \code{DataFrame} with an attribute \code{err} and columns
\itemize{
  \item{\code{seqs}}{: a \code{DNAStringSet} object.}
  \item{\code{is_foreground}}{ that indicates if a sequence is in the
                              foreground group.}
  \item{\code{gc_frac}}{: the fraction of G+C bases per sequence.}
  \item{\code{gc_bin}}{: the GC bin for each sequence.}
  \item{\code{gc_weight}}{: the sequence weight to adjust for GC
    differences between foreground and background sequences.}
  \item{\code{kmer_weight}}{: the sequence weight to adjust for k-mer
    differences between foreground and background sequences.}
}}
}
\value{
a \code{DataFrame} of the same dimensions as the input \code{df},
  with the columns \code{gc_frac}, \code{gc_bin} and \code{gc_weight}
  filled in with the sequence GC content, GC bins they were
  assigned to, and the weight to correct for GC differences between
  foreground and background sequences, respecitvely.
}
\description{
In this implementation we still follow HOMER. But this can be
  re-implemented in a more continuous fashion. Each sequence is put in a
  specific GC bin depending on its GC content. Then, for each GC bin, the
  number of foreGround and background sequences in that bin is calculated.
  Weights are calculated for the background sequences in bin i as follows:
  weight_i = (number_fg_seqs_i / number_bg_seqs_i) * (number_bg_seqs_total /
  number_fg_seqs_total)
}
