% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{clusterKmers}
\alias{clusterKmers}
\title{Cluster k-mers}
\usage{
clusterKmers(
  x,
  allowReverseComplement = FALSE,
  nKmers = NULL,
  maxShift = NULL,
  minSim = NULL
)
}
\arguments{
\item{x}{Enriched k-mers, either a \code{character} vector or the return
value of \code{\link{getKmerFreq}}, from which enriched k-mers can be
extracted.}

\item{allowReverseComplement}{A \code{logical} scalar. If \code{TRUE}, use
the reverse complement of a k-mer if it has a higher toal similarity to all
other k-mers than the original k-mer. In that case, the resulting k-mer is
reported with the suffix \code{_rc} in the results, and the
reverse-complement k-mer sequence is used to calculate pairwise distances.}

\item{nKmers}{A \code{numeric} scalar. If \code{x} is a \code{list} as
returned by \code{\link{getKmerFreq}}, the number of top enriched k-mers to
use ordered by FDR. If \code{NULL} (the default), it will use \code{nKmers
<- max(10, sum(x$FDR < 0.05))}. \code{nKmers} is ignored if \code{x} is a
character vector.}

\item{maxShift}{A \code{numeric} scalar with the maximal number of shifts to
perform when calculating k-mer distances. If \code{NULL} (the default), it
will use \code{maxShift <- kmerLen - 2}.}

\item{minSim}{A \code{numeric} scalar with the minimal k-mer similarity to
link two k-mer nodes in the graph by an edge. If \code{NULL}, it will use
\code{minSim <- kmerLen - maxShift + 1}.}
}
\value{
A named \code{numeric} vector, with k-mer as names and values
  indicating the k-mer cluster memberships.
}
\description{
Given enriched k-mers (typically the result of
  \code{\link{getKmerFreq}}), cluster k-mers into groups of similar k-mers
  that are likely to originate from the same motif.
}
\details{
The clustering is performed as follows: First, all pairwise k-mer
  distances for all possible shifts (defined by \code{maxShift}) are
  calculated, defined as the Hamming distance of the overlapping substring
  plus the number of shifts. For each k-mer pair, the minimal distance is
  retained.  If \code{allowReverseComplement == TRUE}, repeat this procedure
  to compare each k-mer to all reverse-complemented k-mers, and replace it
  with the reverse-complemented version if this yields a lower sum of
  pairwise distances. The resulting distance matrix is then converted into a
  similarity matrix by subtracting it from \code{kmerLen} (the length of the
  k-mers). Element less than \code{minSim} are set to zero, and the matrix is
  used as an adjacency matrix to construct a k-mer graph. The clusters are
  identified as the connected components in this graph.
}
