% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_HOMER.R
\name{clusterPWMs}
\alias{clusterPWMs}
\title{Calculate similarity matrix of motifs.}
\usage{
clusterPWMs(motifFile, homerdir, outfile)
}
\arguments{
\item{motifFile}{A file with HOMER formatted PWMs as input for compareMotifs.pl.}

\item{homerdir}{Path to the HOMER binary directory.}
}
\value{
A matrix of Pearson correlations for each pairwise comparison of
    motifs in motifFile.
}
\description{
Run the HOMER script compareMotifs.pl (with default options) to get a similarity matrix
    of all motifs. For details, see the HOMER documentation.
}
