% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{convertKmersToMotifs}
\alias{convertKmersToMotifs}
\title{Transform k-mer enrichments to motif enrichments.}
\usage{
convertKmersToMotifs(x, m, BPPARAM = SerialParam(), verbose = FALSE)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with the
results of a k-mer enrichment analysis (typically generated by
\code{\link{calcBinnedKmerEnr}}).}

\item{m}{Either a \code{\link[TFBSTools]{PFMatrixList}}, or a character
scalar with a file containing motifs in HOMER format (loaded into a
\code{\link[TFBSTools]{PFMatrixList}} by
\code{\link{homerToPFMatrixList}}).}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}

\item{verbose}{A logical scalar. If \code{TRUE}, report on progress.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}}
    with \code{length(m)} rows (motifs) and \code{ncol(x)} columns (bins).
}
\description{
Using a set of know motifs and the result of a k-mer enrichment
  analysis (k-mers by bins), transform the results into motif enrichments
  (motifs by bins).
}
\examples{
seqs <- Biostrings::DNAStringSet(c("GCATGCATGC", "CATGCGCATG"))
bins <- factor(1:2)
m <- rbind(A = c(2, 0, 0),
           C = c(1, 1, 0),
           G = c(0, 0, 3),
           T = c(0, 2, 0))
pfms <- TFBSTools::PFMatrixList(
    TFBSTools::PFMatrix(name = "m1", profileMatrix = m),
    TFBSTools::PFMatrix(name = "m2", profileMatrix = m[, c(2,1,3)])
)
se1 <- calcBinnedKmerEnr(seqs = seqs, bins = bins, kmerLen = 3)
convertKmersToMotifs(se1, pfms)

}
\seealso{
\code{\link{calcBinnedKmerEnr}} for performing a k-mer enrichment
  analysis, \code{\link[BiocParallel]{bplapply}} used for parallelization.
}
