% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_kmer_pairs}
\alias{count_kmer_pairs}
\title{Count k-mer pairs in sequences}
\usage{
count_kmer_pairs(x, k = 6L, n = 5L, zoops = FALSE)
}
\arguments{
\item{x}{DNAStringSet with sequences.}

\item{k}{An integer scalar with the length of sequences.}

\item{n}{An integer scalar defining the maximum downstream distance of
second k-mers, relative to the start position of the first k-mer.}

\item{zoops}{(not implemented yet)}
}
\value{
A numeric matrix with observed k-mer pairs counts.
}
\description{
Over a set of sequences, for each k-mer a, count occurences of all k-mers b
within a defined distance downstream of the start of a.
}
\examples{
count_kmer_pairs(DNAStringSet(c("AACCGGTT")), k = 2, n = 1)

}
