% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countKmerPairsSelected}
\alias{countKmerPairsSelected}
\title{Count selected k-mer pairs in sequences}
\usage{
countKmerPairsSelected(x, kmers, n = 5L, zoops = FALSE)
}
\arguments{
\item{x}{DNAStringSet with sequences.}

\item{kmers}{DNAStringSet with k-mers.}

\item{n}{An integer scalar defining the maximum downstream distance of
second k-mers, relative to the start position of the first k-mer.}

\item{zoops}{A logical scalar. If TRUE, count each observed k-mer pair
only once per sequence.}
}
\value{
A numeric matrix with observed k-mer pairs counts.
}
\description{
Over a set of sequences, for each k-mer a in a given set of k-mers A, count
occurences of all k-mers b in A within a defined distance downstream of the
start of a.
}
\examples{
countKmerPairs(Biostrings::DNAStringSet(c("AACCGGTT")), k = 2, n = 1)

}
\seealso{
\code{\link{countKmerPairs}} for counting co-occurrences
    amongst all k-mers of a given length. This is more efficient than
    \code{countKmerPairsSelected} and setting \code{kmers} to all
    possible k-mers of a given length.
}
