% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assertColor}
\alias{.assertColor}
\title{Utility function to check validity of R colors.}
\usage{
.assertColor(x, ...)
}
\arguments{
\item{x}{character vector of colors.}

\item{...}{additional arguments to \code{.assertVector}}
}
\description{
This function checks if the provided characters (e.g. arguments to other
functions) are valid colors in R.
}
