% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{.cons2matrix}
\alias{.cons2matrix}
\title{Create matrix from consensus sequence}
\usage{
.cons2matrix(x, n = 100L)
}
\arguments{
\item{x}{Character scalar with the motif the consensus sequence.}

\item{n}{Integer scalar giving the columns sums in the constructed
matrix (number of observed bases at each position).}
}
\description{
Given a nucleotide sequence of A,C,G,T letter corresponding
  to a motif's consensus string, construct a positional frequency matrix.
  This matrix can for example be used as the \code{profileMatrix} argument
  in the constructor for a \code{TFBSTools::PFMatrix} object.
}
\keyword{internal}
