% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{extractOverlappingKmerFrequecies}
\alias{extractOverlappingKmerFrequecies}
\title{Match a set of kmers to input sequences and determine frequencies of overlapping matches.}
\usage{
extractOverlappingKmerFrequecies(seqs, x, Ncpu = 1L)
}
\arguments{
\item{seqs}{Set of sequences, either a \code{character} vector or a
\code{\link{DNAStringSet}}.}

\item{x}{A \code{character} vector of enriched kmers.}

\item{Ncpu}{The number of CPU cores to used. \code{\link[parallel]{mclapply}} is used.}
}
\description{
Using a set of kmers, search input sequences for matches and retrieve
   the frequencies of sequences overlapping with 1 or more overlapping kmers.
}
\seealso{
\code{\link{kmerEnrichments}} for performing a k-mer enrichment
  analysis, \code{\link[parallel]{mclapply}} for how parallelization is done.
}
