% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{extractOverlappingKmerFrequencies}
\alias{extractOverlappingKmerFrequencies}
\title{Match a set of k-mers to input sequences and determine frequencies of
    overlapping matches.}
\usage{
extractOverlappingKmerFrequencies(
  seqs,
  x,
  includeRevComp = TRUE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{seqs}{\code{\link{DNAStringSet}} with sequences to search.}

\item{x}{A \code{character} vector of k-mers to search in \code{seqs}.}

\item{includeRevComp}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE}
(default), \code{seqs} will not only be searched for matches to \code{x},
but also for matches to the reverse-complemented k-mers in \code{x}.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}
}
\value{
A named integer vector with the number of observed occurrences of
  sequences overlapping enriched k-mers. The sequences are given as the
  names, and the elements are sorted by decreasing frequency.
}
\description{
Using a set of k-mers, search input sequences for matches and
    retrieve the frequencies of sequences overlapping with any of the k-mers.
    Overlapping k-mer matches will result in extracted sequences that are
    longer than the input k-mers.
}
\seealso{
\code{\link{calcBinnedKmerEnr}} for performing a k-mer enrichment
    analysis, \code{\link[BiocParallel]{bplapply}} used for parallelization.
}
