% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{filter_seqs}
\alias{filter_seqs}
\title{Filter Bad Sequences}
\usage{
filter_seqs(
  df,
  frac = 0.7,
  min_length = 5L,
  max_length = 100000L,
  verbose = TRUE
)
}
\arguments{
\item{df}{a \code{DataFrame} with an attribute \code{err} and columns
\itemize{
  \item{\code{seqs}}{: a \code{DNAStringSet} object.}
  \item{\code{is_foreground}}{ that indicates if a sequence is in the
                              foreground group.}
  \item{\code{gc_frac}}{: the fraction of G+C bases per sequence.}
  \item{\code{gc_bin}}{: the GC bin for each sequence.}
  \item{\code{gc_weight}}{: the sequence weight to adjust for GC
    differences between foreground and background sequences.}
  \item{\code{kmer_weight}}{: the sequence weight to adjust for k-mer
    differences between foreground and background sequences.}
}}

\item{frac}{A numeric scalar with the maximal fraction of N bases allowed in
a sequence (defaults to 0.7).}

\item{min_length}{The minimum sequence length (default from Homer). Sequences shorter than this will be filtered out.}

\item{max_length}{The maximum sequence length (default from Homer). Sequences bigger than this will be filtered out.}

\item{verbose}{A logical scalar. If \code{TRUE}, report on filtering.}
}
\value{
the filtered \code{df}.
}
\description{
We filter sequences similarly to how HOMER (version 4.11) does it. Namely,
  sequences with more than 70 percent (default) N are removed.
}
\details{
The function follows the \code{removePoorSeq.pl} function from Homer.
}
