% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{filterDirGraph}
\alias{filterDirGraph}
\title{Filter directed graph}
\usage{
filterDirGraph(g, edge_weight_thr)
}
\arguments{
\item{g}{A directed k-mer graph (e.g., from \code{buildDirGraphFromKmers})}

\item{edge_weight_thr}{An edge weight threshold. Edges with weight below 
this threshold will be removed.}
}
\description{
Filter directed graph, by only retaining edges with a 
weight exceeding a given threshold.
}
