% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{filterSeqs}
\alias{filterSeqs}
\title{Filter Bad Sequences}
\usage{
filterSeqs(seqs = NULL, frac = 0.7)
}
\arguments{
\item{seqs}{object of class \code{DNAStringSet} containing all the sequences (foreGround and backGround)}

\item{frac}{the fraction of Ns allowed in a sequence (default set to 0.7)}
}
\value{
object of class \code{DNAStringSet} containing the sequences that passed the filter.
}
\description{
We filter sequences similarly to how HOMER does it. Namely, sequences with more
  than 70% (default) N are removed.
}
\details{
Functions folowing/inspired by HOMER to do motif enrichment analysis
}
