% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{get_motif_hits_in_ZOOPS_mode}
\alias{get_motif_hits_in_ZOOPS_mode}
\title{Get Motif Hits}
\usage{
get_motif_hits_in_ZOOPS_mode(
  df,
  pwmL,
  method = c("homer2"),
  homerfile = findHomer("homer2"),
  min.score = 10L,
  Ncpu = 1L,
  verbose = TRUE
)
}
\arguments{
\item{df}{DNAStringSet object with sequences to analyze}

\item{pwmL}{PWMatrixList object with PWMs of motifs}

\item{method}{the \code{method} parameter in the \code{findMotifHits} function}

\item{min.score}{the \code{min.score} parameter in the \code{findMotifHits} function}

\item{Ncpu}{number of CPUs to use}

\item{verbose}{A logical scalar. If \code{TRUE}, describe motif matrix being created}
}
\value{
a matrix where the rows are the sequences and the columns the motifs. This matrix
  consists of 0 and 1 entries for a motif hit or not, respectively.
}
\description{
Get Motif Hits
}
\details{
TODO: should the returned matrix be a sparse matrix? and should we run this once for all
  sequences in all bins (when doing motif enrichment across bins)?
}
