% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{getGCweight}
\alias{getGCweight}
\title{Get GC weights for background regions}
\usage{
getGCweight(foreGround_seq = NULL, backGround_seq = NULL)
}
\arguments{
\item{foreGround_seq}{object of class \code{DNAStringSet} containing all foreGround sequences}

\item{backGround_seq}{object of class \code{DNAStringSet} containing all backGround sequences}
}
\value{
a list containing:
  \itemize{
    \item{sequenceWeights}{: a \code{dataframe} comtaining the sequence GC content, GC bins
      they were assigned to, and the weight to correct for GC differences between foreGround
      and backGround sequences}
    \item{sequenceNucleotides}{: a \code{DNAStringSet} object containing the raw sequences}
  }
}
\description{
In this implementation we still follow HOMER. But this can be re-implemented in 
  a more continuous fashion. Each sequence is put in a specific GC bin depending on its GC content. 
  Then, for each GC bin, the number of foreGround and backGround sequences in that bin is calculated.
  Weights are calculated for the backGround sequences in bin i as follows:
  weight_i = (number_fg_seqs_i/number_bg_seqs_i)*(number_bg_seqs_total/number_fg_seqs_total)
}
