% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{getMotifsFromDirGraph}
\alias{getMotifsFromDirGraph}
\title{Get putative motifs from directed graph}
\usage{
getMotifsFromDirGraph(g, edge_weight_thr)
}
\arguments{
\item{g}{A directed k-mer graph (e.g., from \code{buildDirGraphFromKmers})}

\item{edge_weight_thr}{An edge weight threshold. Edges with weight below 
this threshold will be removed.}
}
\description{
Get putative motifs from a directed graph, by only retaining edges with a 
weight exceeding a given threshold.
}
