% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_selection.R
\name{glmnet.randomized_lasso}
\alias{glmnet.randomized_lasso}
\title{Randomized Lasso}
\usage{
glmnet.randomized_lasso(x, y, q, weakness = 1, type = c("conservative",
  "anticonservative"), ...)
}
\arguments{
\item{x}{the predictor matrix}

\item{y}{the response vector}

\item{q}{the average number fo selected variables}

\item{weakness}{the parameter essential for the randomized lasso.}

\item{type}{}
}
\value{
the regression output
}
\description{
This function perform randomized lasso using the \code{glmnet} package. The 
function present in the \code{stabs} package that runs the lasso version was adapted for
the randomized lasso here. Randmized lasso stability selection uses this function repeatedly
to select predictors.
}
\author{
Dania Machlab
}
