% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_HOMER.R
\name{homerToPFMatrixList}
\alias{homerToPFMatrixList}
\title{Read a HOMER motif file and create a TFMatrixList.}
\usage{
homerToPFMatrixList(filename, n = 100L)
}
\arguments{
\item{filename}{Name of the input file with HOMER-formatted motifs.}

\item{n}{The number of observations (multiplied with base frequencies to
create the number of observed bases at each position).}
}
\value{
A \code{\link[TFBSTools]{PFMatrixList}} with motifs from the file.
}
\description{
Read motifs from a file in HOMER format and create
    a TFMatrixList from them.
}
\seealso{
\code{\link{dumpJaspar}} for writing motifs from a Jaspar database
    package into a file in HOMER format.
}
