% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{iterate_norm_for_kmer_comp}
\alias{iterate_norm_for_kmer_comp}
\title{Adjust for k-mer composition (multiple iterations)}
\usage{
iterate_norm_for_kmer_comp(df, max_autonorm_iters = 160L)
}
\arguments{
\item{df}{a \code{DataFrame} as returned by \code{calculate_GC_weight}}

\item{max_autonorm_iters}{An integer scalar giving the maximum number if
times to run \code{norm_for_kmer_comp}. the default is set to 160 (as in
\code{HOMER}).}
}
\value{
a list containing: \itemize{ \item{sequenceWeights}{: a
  \code{dataframe} containing the sequence GC content, GC bins they were
  assigned to, the weight to correct for GC differences between foreGround
  and background sequences, the weight to adjust for kmer composition, and
  the the error term} \item{sequenceNucleotides}{: a \code{DNAStringSet}
  object containing the raw sequences} }
}
\description{
Here we run `norm_for_kmer_comp` multiple times to converge to
  the final weights that will be used to correct the background
  sequences for k-mer composition differences compared to the foreground. We
  closely follow \code{HOMER}'s \code{normalizeSequence()} function found in
  \cod{Motif2.cpp}. Note that \code{HOMER} runs the
  \code{normalizeSequence()} one last time after going through all iterations
  or reaching a low error, which we do not do here.
}
