% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{kmerEnrichments}
\alias{kmerEnrichments}
\title{Run a k-mer enrichment analysis.}
\usage{
kmerEnrichments(x, b, genomepkg = NULL, kmerLen = 4, MMorder = 2,
  pseudoCount = 1, Ncpu = 2L, verbose = TRUE)
}
\arguments{
\item{x}{A \code{character} vector, \code{\link[Biostrings]{DNAStringSet}} or
a \code{\link[GenomicRanges]{GRanges}} object with the sequences to analyze.}

\item{b}{A vector of the same length as \code{x} that groups its elements
into bins (typically a factor, such as the one returned by
\code{\link{bin}}).}

\item{genomepkg}{Only used if \code{x} is a \code{GRanges} object: A
\code{character} scalar with the name of a \code{BSgenome} package from
which to extract the sequences.}

\item{kmerLen}{A \code{numeric} scalar giving the k-mer length.}

\item{MMorder}{A \code{numeric} scalar giving the order of the Markov model
used to calculate the expected frequencies.}

\item{pseudoCount}{A \code{numeric} scalar - will be added to the observed
counts for each k-mer to avoid zero values.}

\item{Ncpu}{Number of parallel threads to use.}

\item{verbose}{A logical scalar. If \code{TRUE}, report on progress.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} \code{y}
  with \describe{
    \item{assays(y)}{containing the four components \code{p}, \code{FDR},
       \code{enr} and \code{log2enr}), each a k-mer (rows) by bin (columns)
       matrix with raw -log10 P values, -log10 false discovery rates and
       k-mer enrichments as Pearson residuals (\code{enr}) and as log2 ratios
       (\code{log2enr}).}
    \item{rowData(x)}{containing information about the k-mers.}
    \item{colData(x)}{containing information about the bins.}
    \item{metaData(x)}{containing meta data on the object (e.g. parameter values).}
  }
}
\description{
Given a set of sequences and corresponding bins, identify
  enriched k-mers (n-grams) in each bin. The sequences can be given either
  directly or as genomic coordinates.
}
\seealso{
\code{\link{getKmerFreq}} used to calculate k-mer enrichments;
  \code{\link[BSgenome]{getSeq,BSgenome-method}} which is used to extract
  sequences from \code{genomepkg} if \code{x} is a \code{GRanges} object;
  \code{\link[parallel]{mclapply}} that is used for parallelization;
  \code{\link{bin}} for binning of regions
}
