% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\docType{package}
\name{lisa-package}
\alias{lisa}
\alias{lisa-package}
\title{Lisa - motif enrichments and more.}
\description{
\pkg{lisa} is a collection of tools that simplify motif enrichment analyses
in genomic regions of interest.
}
\details{
She makes use of her father Homer (http://homer.ucsd.edu/homer/index.html)
and other algorithms to search for motif hits and look for enriched motifs in
sets of genomic regions, compared to all other regions.

Known motifs can for example be obtained from a collection of transcription
factor binding site specificities, such as \pkg{JASPAR2018}.
}
\author{
\strong{Maintainer}: Michael Stadler \email{michael.stadler@fmi.ch}

Authors:
\itemize{
  \item Dania Machlab \email{dania.machlab@fmi.ch}
  \item Lukas Burger \email{lukas.burger@fmi.ch}
}

}
