% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monaLisa-package.R
\docType{package}
\name{monaLisa-package}
\alias{monaLisa-package}
\alias{monaLisa}
\title{monaLisa - MOtif aNAlysis with Lisa.}
\description{
\pkg{monaLisa} is a collection of tools that simplify motif enrichment
analyses in genomic regions of interest.
}
\details{
She makes use of her father Homer (http://homer.ucsd.edu/homer/index.html)
and other algorithms to search for motif hits and look for enriched motifs in
sets of genomic regions, compared to all other regions.

Known motifs can for example be obtained from a collection of transcription
factor binding site specificities, such as \pkg{JASPAR2020}.
}
\author{
Dania Machlab, Lukas Burger and Michael Stadler
}
\keyword{package}
