% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{multipleNormalizeForKmerComposition}
\alias{multipleNormalizeForKmerComposition}
\title{Get Final Sequence Weights}
\usage{
multipleNormalizeForKmerComposition(
  inputList = NULL,
  maxAutoNormIters = 160,
  lastError = 1e+100
)
}
\arguments{
\item{inputList}{a `list` resulting from running `getGCweight`}

\item{maxAutoNormIters}{the maximum number if times to run `normalizeForKmerComposition`. the default
is set to 160 (as in `HOMER`).}

\item{lastError}{teh starting value for the last error. Once `normalizeForKmerComposition` is run, the aim is 
to stop when teh current error is bigger than the last error.}
}
\value{
a list containing:
  \itemize{
    \item{sequenceWeights}{: a \code{dataframe} containing the sequence GC content, GC bins
      they were assigned to, the weight to correct for GC differences between foreGround
      and backGround sequences, the weight to adjust for kmer composition, and the the error term}
    \item{sequenceNucleotides}{: a \code{DNAStringSet} object containing the raw sequences}
  }
}
\description{
Here we run `normalizeForKmerComposition` multiple times to get the final set of weights that 
  will be used to correct the background sequences for kmer composition differences compared to the foreGround.
  We closely follow `HOMER`'s `normalizeSequence()` function found in `Motif2.cpp`. Note that `HOMER`
  runs the `normalizeSequence()` one last time after going through all iterations or reaching a low error, 
  which we do not do here.
}
