% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{norm_for_kmer_comp}
\alias{norm_for_kmer_comp}
\title{Adjust for k-mer composition (single iteration)}
\usage{
norm_for_kmer_comp(
  kmer_freq,
  g_oligos,
  kmer_seq_rc,
  kmer_weight,
  is_foreground,
  minimum_seq_weight = 0.001,
  maximum_seq_weight = 1000
)
}
\arguments{
\item{kmer_freq}{a \code{list} with of matrices. The matrix at index \code{i}
contains the counts of k-mers of length \code{i} (columns) for each
sequence (rows).}

\item{g_oligos}{a \code{list} of \code{numeric} vectors; the element at index
\code{i} contains the number of good (non-N-containing) k-mers of length
\code{i} for each sequence.}

\item{kmer_seq_rc}{a \code{list} of character vectors; the element at index
\code{i} contains the reverse complement sequences of all k-mers of length
\code{i}.}

\item{kmer_weight}{a \code{numeric} vector with sequence weights adjusting
for k-mer composition at the beginning of the iteration.}

\item{is_foreground}{logical vector of the same length as \code{seqs}.
\code{TRUE} indicates that the sequence corresponds to the foreground set,
\code{FALSE} indicates a background set sequence.}

\item{minimum_seq_weight}{Numeric scalar greater than zero giving the
minimal weight of a sequence. The default value (0.001) was also used by
\code{HOMER} (HOMER_MINIMUM_SEQ_WEIGHT constant in Motif2.h).}

\item{maximum_seq_weight}{Numeric scalar greater than zero giving the
maximal weight of a sequence. The default value (1000) was also used by
\code{HOMER} (1 / HOMER_MINIMUM_SEQ_WEIGHT constant in Motif2.h).}
}
\value{
a \code{DataFrame} of the same dimensions as the input \code{df},
  with the weight to adjust for k-mer composition and the current error
  stored in the column \code{kmer_weight} and the attribute \code{err}.
}
\description{
Here we correct the background sequence weights, adjusting for
  k-mer composition compared to the foreground sequences. This function
  implements a single iteration, and is called iteratively by
  \code{iterate_norm_for_kmer_comp} to get to the final set of adjusted
  weights, which will be the result of adjusting for GC and k-mer
  composition. We closely follow HOMER's \code{normalizeSequenceIteration()}
  function found in \code{Motif2.cpp}.
}
