% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{normalizeForKmerComposition}
\alias{normalizeForKmerComposition}
\title{Adjust for Kmer Composition}
\usage{
normalizeForKmerComposition(inputList = NULL, maxKmerSize = 3)
}
\arguments{
\item{inputList}{a `list`.}

\item{maxKmerSize}{the maximum kmer size to consider. The default is set to 3 (like in `HOMER`), 
meaning that kmers of size 1,2 and 3 are considered.}
}
\value{
a list containing:
  \itemize{
    \item{sequenceWeights}{: a \code{dataframe} containing the sequence GC content, GC bins
      they were assigned to, the weight to correct for GC differences between foreGround
      and backGround sequences, the weight to adjust for kmer composition, and the the error term}
    \item{sequenceNucleotides}{: a \code{DNAStringSet} object containing the raw sequences}
  }
}
\description{
Here we correct the backGround sequence weights, adjusting for kmer composition
  compared to the foreGroud sequences. This function implements one run, and it will be iteratively 
  called multiple times by another function to get to the final set of adjusted weights. These 
  weights will the result of adjusting for GC and kmer composition. We closely follow `HOMER`'s 
  `normalizeSequenceIteration()` function found in `Motif2.cpp`.
}
