% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_stabilityPaths}
\alias{plot_stabilityPaths}
\title{Plot Stability Paths}
\usage{
plot_stabilityPaths(stabs_object, cols = NULL, lty = 1, ...)
}
\arguments{
\item{stabs_object}{the \code{stabs} object resulting from stability selection.}

\item{cols}{color vector for the varaiables from the predictor matrix. By default, it's set to NULL
and the function colors the variables by whether or not they were selected.}
}
\value{
plot of stability paths.
}
\description{
Plot the stability paths of each variable (predictor), showing the selection probability
as a function of the lambda parameter.
}
