% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBinScatter}
\alias{plotBinScatter}
\title{Scatter plot (xy-plot) of binned elements.}
\usage{
plotBinScatter(
  x,
  y,
  b,
  cols = getColsByBin(b),
  xlab = deparse(substitute(x, env = as.environment(-1))),
  ylab = deparse(substitute(y, env = as.environment(-1))),
  main = "",
  legend = "topright",
  legend.cex = 1,
  ...
)
}
\arguments{
\item{x}{A numerical vector with x values.}

\item{y}{A numerical vector with y values (the values used for binning).}

\item{b}{A factor that groups elements of \code{x,y} into bins (typically
the output of \code{\link{bin}(y)}).}

\item{cols}{A color vector (will be computed based on \code{b} by default
using \code{\link{getColsByBin}(b)}).}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{main}{Main title.}

\item{legend}{If not \code{NULL}, draw a legend with binning information
(will be passed to \code{legend(x=legend)} to control legend position).}

\item{legend.cex}{A scalar that controls the text size in the legend relative
to the current \code{par("cex")} (see \code{\link{legend}}).}

\item{...}{Further arguments passed to \code{plot(x, y, ...)}.}
}
\value{
\code{TRUE} (invisibly).
}
\description{
Plot a scatter (xy-plot) of binned elements with binning
    information.
}
\examples{
set.seed(1)
x <- rnorm(100)
y <- rnorm(100)
b <- bin(y, "equalN", nElements = 10)
plotBinScatter(x, y, b)

}
\seealso{
\code{\link{bin}}, \code{\link{getColsByBin}}
}
