% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotStabilityPaths}
\alias{plotStabilityPaths}
\title{Plot Stability Paths}
\usage{
plotStabilityPaths(
  se,
  selProbMin = metadata(se)$stabsel.params.cutoff,
  selColor = "cadetblue",
  notSelColor = "grey",
  selProbCutoffColor = "firebrick",
  linewidth = 0.5,
  alpha = 1,
  ylim = c(0, 1)
)
}
\arguments{
\item{se}{the \code{SummarizedExperiment} object resulting from stability
selection, by running \code{\link[monaLisa]{randLassoStabSel}}.}

\item{selProbMin}{A numerical scalar in [0,1]. Predictors with a selection
probability greater than \code{selProbMin} are shown as colored lines. The
color is defined by the \code{col} argument.}

\item{selColor}{color for the selected predictors which have a selection
probability greater than \code{selProbMin}.}

\item{notSelColor}{color for the rest of the (un-selected) predictors.}

\item{selProbCutoffColor}{color for the line depicting the selection 
probability cutoff.}

\item{linewidth}{line width.}

\item{alpha}{line transparency of the stability paths.}

\item{ylim}{limits for y-axis.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot the stability paths of each variable (predictor),
  showing the selection probability as a function of the regularization step.
}
\examples{
## create data set
Y <- rnorm(n = 500, mean = 2, sd = 1)
X <- matrix(data = NA, nrow = length(Y), ncol = 50)
for (i in seq_len(ncol(X))) {
  X[ ,i] <- runif(n = 500, min = 0, max = 3)
}
s_cols <- sample(x = seq_len(ncol(X)), size = 10,
  replace = FALSE)
for (i in seq_along(s_cols)) {
  X[ ,s_cols[i]] <- X[ ,s_cols[i]] + Y
}

## reproducible randLassoStabSel() with 1 core
set.seed(123)
ss <- randLassoStabSel(x = X, y = Y)
plotStabilityPaths(ss)

}
\seealso{
\code{\link[stabs]{stabsel}}
}
