% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_selection.R
\name{randomized_stabsel}
\alias{randomized_stabsel}
\title{Randomized Lasso Stability Selection}
\usage{
randomized_stabsel(x = x, y = y, weakness = 0.8, cutoff = 0.8,
  PFER = 2, ...)
}
\arguments{
\item{x}{the predictor matrix}

\item{y}{the response vector}

\item{weakness}{the parameter essential for the randomized lasso.}

\item{cutoff}{value between 0 and 1 which is the cutoff for the selection probability. Any variable
with a selection probability that is higher than the set cutoff will be selected.}

\item{PFER}{the absolute number of false positives that we allow for in the final list of 
selected variables. This comes from the Meinshausen and Buehlmann paper.}
}
\value{
a \code{stabsel} object
}
\description{
This function runs randomized lasso stability selection as presented by Meinshausen (2010) 
and Buehlmann and with the improved error bounds introduced by Shah and Samworth (2013). The function 
uses the \code{stabsel} function from the \code{stabs} package, but using the randomized lasso version.
}
