% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_HOMER.R
\name{runHomer}
\alias{runHomer}
\title{Prepare and run HOMER motif enrichment analysis.}
\usage{
runHomer(
  gr,
  b,
  genomedir,
  outdir,
  motifFile,
  homerfile = findHomer(),
  regionsize = "given",
  Ncpu = 2L,
  verbose = FALSE
)
}
\arguments{
\item{gr}{A \code{GRanges} object (or an object that can be coerced to one)
with the genomic regions to analyze.}

\item{b}{A vector of the same length as \code{gr} that groups its elements
into bins (typically a factor, such as the one returned by \code{\link{bin}}).}

\item{genomedir}{Directory containing sequence files in Fasta format (one per chromosome).}

\item{outdir}{A path specifying the folder into which the output files will be written.}

\item{motifFile}{A file with HOMER formatted PWMs to be used in the enrichment analysis.}

\item{homerfile}{Path and file name of the \code{findMotifsGenome.pl} HOMER script.}

\item{regionsize}{The peak size to use in HOMER (\code{"given"} keeps the coordinate
region, an integer value will keep only that many bases in the region center).}

\item{Ncpu}{Number of parallel threads that HOMER can use.}

\item{verbose}{A logical scalar. If \code{TRUE}, print progress messages.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} \code{x}
  with
  \describe{
    \item{assays(x)}{containing the four components \code{p}, \code{FDR},
        \code{enr} and \code{log2enr}), each a motif (rows) by bin (columns)
        matrix with raw -log10 P values, -log10 false discovery rates and motif
        enrichments as Pearson residuals (\code{enr}) and as log2 ratios
        (\code{log2enr}).}
    \item{rowData(x)}{containing information about the motifs.}
    \item{colData(x)}{containing information about the bins.}
    \item{metaData(x)}{containing meta data on the object (e.g. parameter
        values and distances between pairs of motifs).}
  }
}
\description{
Run complete HOMER motif enrichment analysis, consisting of
    calls to \code{\link{prepareHomer}}, \code{\link[base]{system}} and
    \code{\link{parseHomerOutput}}.
}
\seealso{
The functions that are wrapped: \code{\link{prepareHomer}},
    \code{\link[base]{system}} and \code{\link{parseHomerOutput}},
    \code{\link{bin}} for binning of regions
}
