% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{runMonaLisa}
\alias{runMonaLisa}
\title{Do Motif Enrichment Analysis with `monaLisa` (for simple fg vs bg situation, this wil lbe changed for bins later)}
\usage{
runMonaLisa(seqs = NULL, foreGround = NULL)
}
\description{
Do Motif Enrichment Analysis with `monaLisa` (for simple fg vs bg situation, this wil lbe changed for bins later)
}
