% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{summarize_binned_enrichment_results}
\alias{summarize_binned_enrichment_results}
\title{Summarize Binned Enrichmetn results}
\usage{
summarize_binned_enrichment_results(se, pseudo = 0.01)
}
\arguments{
\item{se}{a \code{SummarizedExperiment} object resulting from running the \code{get_binned_motif_enrichment} function.}

\item{pseudo}{the pseudo count used when calculating the log2 enrichment for foreground over background.}
}
\value{
A \code{SummarizedExperiment} object where the rows are the bins. The assays represent summarized results: An 
  assay called "log2_enrichment" that contains the log2(enrichment) values, and an assay called "log_p_value" that 
  contains the log p-values.
}
\description{
This function summarizes the output from \code{get_binned_motif_enrichment} to 
  produce matrices that depict the log enrichment and log p-values for each motif across
  the defined bins.
}
