% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_selection.R
\name{TFBS_perSeqName}
\alias{TFBS_perSeqName}
\title{TFBS Matrix}
\usage{
TFBS_perSeqName(TFBS_gr, subject_gr, PWMs)
}
\arguments{
\item{TFBS_gr}{the output of \code{findMotifHits} contaiting the TF binding locations across specified genomic regions.}

\item{subject_gr}{a \code{GRanges} object showing the positions of the geominc regions That have been scanned for the TFs. 
This corresponds to the \code{GRanges} \code{subject} parameter used in the \code{findMorifHits} function.}

\item{PWMs}{\code{PWMatrixList} or \code{PWMatrix} object of the used TFs. This corresponds to the \code{query}
parameter used in the \code{findMorifHits} function.}
}
\value{
a matrix containing the number of binding sites each TF has across the genomic regions.
}
\description{
This function takes in a \code{GRanges} that is the output of \code{findMotifHits} and returns a 
matrix contaiting the number of transcription factor bisnding sites (TFBS) per motif across a given set of genomic ranges.
}
\details{
More details to come
}
\author{
Dania Machlab
}
