% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{calcBinnedKmerEnr}
\alias{calcBinnedKmerEnr}
\title{Calculate k-mer enrichment in bins of sequences.}
\usage{
calcBinnedKmerEnr(
  x,
  b,
  genomepkg = NULL,
  kmerLen = 5,
  background = c("other", "model"),
  MMorder = 1,
  zoops = TRUE,
  p.adjust.method = "BH",
  pseudoCount = 1,
  BPPARAM = SerialParam(),
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{character} vector, \code{\link[Biostrings]{DNAStringSet}} or
a \code{\link[GenomicRanges]{GRanges}} object with the sequences to
analyze.}

\item{b}{A vector of the same length as \code{x} that groups its elements
into bins (typically a factor, such as the one returned by
\code{\link{bin}}).}

\item{genomepkg}{Only used if \code{x} is a \code{GRanges} object: A
\code{character} scalar with the name of a \code{BSgenome} package from
which to extract the sequences.}

\item{kmerLen}{A \code{numeric} scalar giving the k-mer length.}

\item{background}{A \code{character} scalar. If \code{"other"} (the default),
the enrichments in a bin are calculated compared to the sequences in all
other bins. If \code{"model"}, the enrichments are relative to the expected
frequencies obtained from a Markov model of order \code{MMorder}.}

\item{MMorder}{A \code{numeric} scalar giving the order of the Markov model
used to calculate the expected frequencies for \code{background = "model"}.}

\item{zoops}{A \code{logical} scalar. If \code{TRUE} (the default), only one
or zero occurences of a k-mer are considered per sequence. This is helpful
to reduce the impact of simple sequence repeats occurring in few sequences.}

\item{p.adjust.method}{A character scalar selecting the p value adjustment
method (used in \code{\link[stats]{p.adjust}}).}

\item{pseudoCount}{A \code{numeric} scalar - will be added to the observed
counts for each k-mer to avoid zero values.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}

\item{verbose}{A \code{logical} scalar. If \code{TRUE}, report on progress.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object with
  k-mers in rows and bins in columns, containing four assays: \itemize{
  \item{negLog10P}{: -log10 P values}
  \item{negLog10Padj}{: -log10 adjusted P values}
  \item{pearsonResid}{: motif enrichments as Pearson residuals}
  \item{log2enr}{: motif enrichments as log2 ratios}
}.
rowData(x) contains information about the k-mers, colData(x) contains
information about the bins, and metaData(x) contains meta data on the object
and how it was generated (e.g. parameter values).
}
\description{
Given a set of sequences and corresponding bins, identify
  enriched k-mers (n-grams) in each bin. The sequences can be given either
  directly or as genomic coordinates.
}
\examples{
seqs <- DNAStringSet(c("GCATGCATGC", "CTAGCTAGCTG"))
bins <- factor(1:2)
calcBinnedKmerEnr(x = seqs, b = bins, kmerLen = 3)

}
\seealso{
\code{\link{getKmerFreq}} used to calculate k-mer enrichments;
  \code{\link[BSgenome]{getSeq,BSgenome-method}} which is used to extract
  sequences from \code{genomepkg} if \code{x} is a \code{GRanges} object;
  \code{\link[BiocParallel]{bplapply}} that is used for parallelization;
  \code{\link{bin}} for binning of regions
}
