% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{calcBinnedMotifEnrR}
\alias{calcBinnedMotifEnrR}
\title{Binned Motif Enrichment Analysis with \code{monaLisa}}
\usage{
calcBinnedMotifEnrR(
  seqs,
  bins,
  pwmL,
  test = c("binomial", "fisher"),
  maxFracN = 0.7,
  maxKmerSize = 3L,
  min.score = 10,
  matchMethod = "matchPWM",
  pseudocount.log2enr = 8,
  p.adjust.method = "BH",
  BPPARAM = SerialParam(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{seqs}{DNAStringSet object with sequences to test}

\item{bins}{factor of the 
same length and order as \code{seqs}, indicating
  the bin for each sequence. Typically the return value of
  \code{\link[monaLisa]{bin}}.}

\item{pwmL}{PWMatrixList with motifs for which to calculate enrichments.}

\item{test}{A \code{character} scalar specifying the type of enrichment test
to perform. One of \code{"binomial"} (default) or \code{"fisher"}. The
enrichment test is one-sided (enriched in foreground).}

\item{maxFracN}{A numeric scalar with the maximal fraction of N bases allowed
in a sequence (defaults to 0.7). Sequences with higher fractions are
excluded from the analysis.}

\item{maxKmerSize}{the maximum k-mer size to consider, when adjusting
background sequence weights for k-mer composition compared to the
foreground sequences. The default value (3) will correct for mono-, di-
and tri-mer composition.}

\item{min.score}{the minimal score for motif hits, used in
\code{\link[monaLisa]{findMotifHits}}.}

\item{matchMethod}{the method used to scan for motif hits, passed to the
\code{method} parameter in \code{\link[monaLisa]{findMotifHits}}.}

\item{pseudocount.log2enr}{A numerical scalar with the pseudocount to add to
foreground and background counts when calculating log2 motif enrichments}

\item{p.adjust.method}{A character scalar selecting the p value adjustment
method (used in \code{\link[stats]{p.adjust}}).}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}

\item{verbose}{A logical scalar. If \code{TRUE}, print progress messages.}

\item{...}{Additional arguments for  \code{\link[monaLisa]{findMotifHits}}.}
}
\value{
A \code{SummarizedExperiment} object with motifs in rows and bins
  in columns, containing six assays: \itemize{
  \item{negLog10P}{: -log10 P values}
  \item{negLog10Padj}{: -log10 adjusted P values}
  \item{pearsonResid}{: motif enrichments as Pearson residuals}
  \item{log2enr}{: motif enrichments as log2 ratios}
  \item{sumForegroundWgtWithHits}{: Sum of foreground sequence weights
    in a bin that have motif hits}
  \item{sumBackgroundWgtWithHits}{: Sum of background sequence weights
    in a bin that have motif hits}
}
}
\description{
This function performs a motif enrichment analysis on bins of
  sequences. For each bin, the sequences in all other bins are used as
  background.
}
\details{
This function implements a binned motif enrichment analysis. In each
  enrichment analysis, the sequences in a specific bin are used as foreground
  sequences to test for motif enrichment using sequences in the remaining
  bins as background. The logic follows the \code{findMotifsGenome.pl} tool
  from \code{Homer} version 4.11, with \code{-size given -nomotif -mknown}
  and gives very similar results. As in the \code{Homer} tool, sequences are
  weighted to correct for GC and k-mer composition differences between fore-
  and background sets.

  Motif hits are predicted using \code{\link[monaLisa]{findMotifHits}} and
  multiple hits per sequence are counted as just one hit (ZOOPS mode). For
  each motif, the weights of sequences that have a hit are summed separately
  for foreground (\code{sumForegroundWgtWithHits}) and background
  (\code{sumBackgroundWgtWithHits}). The total foreground 
  (\code{totalWgtForeground}) and background (\code{totalWgtBackground})
  sum of sequence weights is also calculated.

  To statistical tests for the calculation of enrichment log p-value are
  available: \code{test = "binomial"} (default) to perform binomial test
  like \code{Homer}, or \code{test = "fisher"} to perform Fisher's exact
  tests, using:
  \itemize{
    \item{binomial}{: \code{pbinom(q = sumForegroundWgtWithHits - 1, size =
      totalWgtForeground, prob = sumBackgroundWgtWithHits / totalWgtBackground,
      lower.tail = FALSE, log.p = TRUE)}}
    \item{fisher}{: \code{fisher.test(x = tab, alternative =
      "greater")}, where \code{tab} is the contingency table with the summed
      weights of sequences in foreground or background sets (rows), and with
      or without a hit for a particular motif (columns).}
  }
}
