% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{convertKmersToMotifs}
\alias{convertKmersToMotifs}
\title{Transform k-mer enrichments to motif enrichments.}
\usage{
convertKmersToMotifs(x, m, BPPARAM = SerialParam(), verbose = TRUE)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with the
results of a k-mer enrichment analysis (typically generated by
\code{\link{kmerEnrichments}}).}

\item{m}{Either a \code{\link[TFBSTools]{PFMatrixList}}, or a character
scalar with a file containing motifs in HOMER format (loaded into a
\code{\link[TFBSTools]{PFMatrixList}} by
\code{\link{homerToPFMatrixList}}).}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}

\item{verbose}{A logical scalar. If \code{TRUE}, report on progress.}
}
\description{
Using a set of know motifs and the result of a k-mer enrichment
  analysis (k-mers by bins), transform the results into motif enrichments
  (motifs by bins).
}
\seealso{
\code{\link{kmerEnrichments}} for performing a k-mer enrichment
  analysis, \code{\link[BiocParallel]{bplapply}} used for parallelization.
}
