% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{.calculateGCweight}
\alias{.calculateGCweight}
\title{Get background sequence weights for GC bins}
\usage{
.calculateGCweight(df, verbose = TRUE)
}
\arguments{
\item{df}{a \code{DataFrame} with sequence information.}

\item{verbose}{A logical scalar. If \code{TRUE}, report on GC weight
calculation.}
}
\value{
a \code{DataFrame} of the same dimensions as the input \code{df},
  with the columns \code{gc_frac}, \code{gc_bin} and \code{gc_weight}
  filled in with the sequence GC content, assigned GC bins and weights to
  correct differences in GC distributions between foreground and background
  sequences.
}
\description{
The logic is based on Homer (version 4.11). All sequences
  binned depending on GC content (hard-coded bin breaks). The numbers of
  foreground and background sequences in each bin are counted, and weights
  for background sequences in bin i are defined as:
  weight_i = (number_fg_seqs_i / number_bg_seqs_i) * (number_bg_seqs_total /
  number_fg_seqs_total)
}
