% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enrichment.R
\name{.checkIfSeqsAreEqualLength}
\alias{.checkIfSeqsAreEqualLength}
\title{Check if elements of `x` are have equal lengths}
\usage{
.checkIfSeqsAreEqualLength(x)
}
\arguments{
\item{x}{an object that implements a \code{width} method, typically a
\code{GRanges} or \code{DNAStringSet} object.}
}
\value{
\code{NULL} (invisibly). The function is called for its side-effect
  of generating a warning if elements of the input are not of equal lengths.
}
\description{
Check if the elements of `x` are all equally long.
  If not, generate a warning.
}
\keyword{internal}
