% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_HOMER.R
\name{dumpJaspar}
\alias{dumpJaspar}
\title{Dump Jaspar motifs into a HOMER motif file.}
\usage{
dumpJaspar(
  filename,
  pkg = "JASPAR2018",
  opts = list(tax_group = "vertebrates"),
  relScoreCutoff = 0.8,
  verbose = TRUE
)
}
\arguments{
\item{filename}{Name of the output file to be created.}

\item{pkg}{Name of the Jaspar package to use (default: \code{JASPAR2018}).}

\item{opts}{a search options list used in \code{\link[TFBSTools]{getMatrixSet}}.}

\item{relScoreCutoff}{Currently ignored. numeric(1) in [0,1] that sets the default motif
log-odds score cutof to relScoreCutoff * maximal score for each PWM
(default: 0.8).}

\item{verbose}{A logical scalar. If \code{TRUE}, print progress messages.}
}
\value{
\code{TRUE} if successful.
}
\description{
Get PWMs from the \code{JASPAR2016} or \code{JASPAR2018} package
    and write them into a HOMER-compatible motif file.
}
\seealso{
\code{\link[TFBSTools]{getMatrixSet}} for details on the argument \code{opts}.
    \code{\link{homerToPFMatrixList}} to read a file with HOMER-formatted
    moitfs into a \code{\link[TFBSTools]{PFMatrixList}}.
}
