% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{extractOverlappingKmerFrequecies}
\alias{extractOverlappingKmerFrequecies}
\title{Match a set of k-mers to input sequences and determine frequencies of overlapping matches.}
\usage{
extractOverlappingKmerFrequecies(seqs, x, BPPARAM = SerialParam())
}
\arguments{
\item{seqs}{Set of sequences, either a \code{character} vector or a
\code{\link{DNAStringSet}}.}

\item{x}{A \code{character} vector of enriched k-mers.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}
}
\value{
A named integer vector with the number of observed occurrences of
  sequences overlapping enriched k-mers. The sequences are given as the
  names, and the elements are sorted by decreasing frequency.
}
\description{
Using a set of k-mers, search input sequences for matches and retrieve
     the frequencies of sequences overlapping with 1 or more overlapping k-mers.
}
\seealso{
\code{\link{kmerEnrichments}} for performing a k-mer enrichment
    analysis, \code{\link[BiocParallel]{bplapply}} used for parallelization.
}
