% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{extractOverlappingKmerFrequecies}
\alias{extractOverlappingKmerFrequecies}
\title{Match a set of kmers to input sequences and determine frequencies of overlapping matches.}
\usage{
extractOverlappingKmerFrequecies(seqs, x, BPPARAM = SerialParam())
}
\arguments{
\item{seqs}{Set of sequences, either a \code{character} vector or a
\code{\link{DNAStringSet}}.}

\item{x}{A \code{character} vector of enriched kmers.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}
}
\description{
Using a set of kmers, search input sequences for matches and retrieve
     the frequencies of sequences overlapping with 1 or more overlapping kmers.
}
\seealso{
\code{\link{kmerEnrichments}} for performing a k-mer enrichment
    analysis, \code{\link[BiocParallel]{bplapply}} used for parallelization.
}
