% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{filterSeqs}
\alias{filterSeqs}
\title{Filter Bad Sequences}
\usage{
filterSeqs(inputList = NULL, frac = 0.7)
}
\arguments{
\item{inputList}{a `list` containing
\itemize{
  \item{sequenceWeights}{: a \code{dataframe} with nrows equal to the number of sequences, with a 
    column called `foreGround` that is 1 if a sequence is in the foreGround group, and 0 otherwise, 
    and a column containing the name of the sequnce.}
  \item{sequenceNucleotides}{: a \code{DNAStringSet} object containing the raw sequences}
}}

\item{frac}{the fraction of Ns allowed in a sequence (default set to 0.7)}
}
\value{
the filtered inputList.
}
\description{
We filter sequences similarly to how HOMER does it. Namely, sequences with more
  than 70% (default) N are removed.
}
