% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_finding.R
\docType{methods}
\name{findMotifHits}
\alias{findMotifHits}
\alias{findMotifHits,ANY-method}
\alias{findMotifHits,character,character-method}
\title{Find motif matches in sequences.}
\usage{
findMotifHits(query, subject, min.score, method = c("homer2", "matchPWM"),
  homerfile = findHomer("homer2"), Ncpu = 1L, ...)

\S4method{findMotifHits}{ANY}(query, subject, min.score, method = c("homer2",
  "matchPWM"), homerfile = findHomer("homer2"), Ncpu = 1L, ...)
}
\arguments{
\item{query}{The motifs to search for, either a
\itemize{
    \item{\code{character(1)}}{ with the path and file name of a motif
          file with PWM in HOMER format (currently only
          supported for \code{method="homer2"})}
    \item{\code{PWMatrix}}{ with a single PWM}
    \item{\code{PWMatrixList}}{ with several PWMs to search for.}
 }}

\item{min.score}{The minimum score for counting a match. Can be given as
a character string containing a percentage (e.g. "85%") of the highest
possible score or as a single number.}

\item{method}{The internal method to use for motif searching. One of
\itemize{
    \item{\code{homer2}}{call to the homer2 binary}
    \item{\code{"matchPWM"}}{using Biostrings::matchPWM}
}}

\item{homerfile}{Path and file name of the \code{homer2} binary.}

\item{Ncpu}{Number of parallel threads (only for \code{method="homer2"}, default: 1).}

\item{...}{Additional arguments for specific methods.}
}
\value{
A \code{GRanges} object with the matches to \code{query} in \code{subject}.
}
\description{
\code{findMotifHits} scans sequences (either provided
as a file, an R object or genomic coordinates)  for matches to
positional weight matrices (provided as a file or as R objects)
}
\examples{
pwm <- TFBSTools::PWMatrix(ID="mypwm", profileMatrix=matrix(rep(c(1,-1,-1,-1),3), nrow=4, dimnames=list(c("A","C","G","T"))))
findMotifHits(pwm, "CCCCCAAACCCCC")
}
\seealso{
\code{\link[Biostrings]{matchPWM}}
}
