% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_finding.R
\docType{methods}
\name{findMotifHits}
\alias{findMotifHits}
\alias{findMotifHits,character,character-method}
\alias{findMotifHits,character,DNAString-method}
\alias{findMotifHits,character,DNAStringSet-method}
\alias{findMotifHits,PWMatrix,character-method}
\alias{findMotifHits,PWMatrix,DNAString-method}
\alias{findMotifHits,PWMatrix,DNAStringSet-method}
\alias{findMotifHits,PWMatrixList,character-method}
\alias{findMotifHits,PWMatrixList,DNAString-method}
\alias{findMotifHits,PWMatrixList,DNAStringSet-method}
\title{Find motif matches in sequences.}
\usage{
findMotifHits(query, subject, ...)

\S4method{findMotifHits}{character,character}(query, subject, min.score,
  method = c("homer2", "matchPWM"), homerfile = findHomer("homer2"),
  Ncpu = 1L, ...)

\S4method{findMotifHits}{character,DNAString}(query, subject, ...)

\S4method{findMotifHits}{character,DNAStringSet}(query, subject, min.score,
  method = c("homer2", "matchPWM"), homerfile = findHomer("homer2"),
  Ncpu = 1L, ...)

\S4method{findMotifHits}{PWMatrix,character}(query, subject, ...)

\S4method{findMotifHits}{PWMatrix,DNAString}(query, subject, ...)

\S4method{findMotifHits}{PWMatrix,DNAStringSet}(query, subject, ...)

\S4method{findMotifHits}{PWMatrixList,character}(query, subject, min.score,
  method = c("homer2", "matchPWM"), homerfile = findHomer("homer2"),
  Ncpu = 1L, ...)

\S4method{findMotifHits}{PWMatrixList,DNAString}(query, subject, ...)

\S4method{findMotifHits}{PWMatrixList,DNAStringSet}(query, subject,
  min.score, method = c("homer2", "matchPWM"),
  homerfile = findHomer("homer2"), Ncpu = 1L, ...)
}
\arguments{
\item{query}{The motifs to search for, either a
\itemize{
    \item{\code{character(1)}}{ with the path and file name of a motif
          file with PWM in HOMER format (currently only
          supported for \code{method="homer2"})}
    \item{\code{PWMatrix}}{ with a single PWM}
    \item{\code{PWMatrixList}}{ with several PWMs to search for.}
 }}

\item{subject}{The sequences to be searched, either a
\itemize{
    \item{\code{character}}{ with the path and file name of a sequence
          file with DNA sequences in FASTA format}
    \item{\code{DNAString}}{ with a single sequence}
    \item{\code{DNAStringSet}}{ with several sequences}
    \item{\code{GRanges}}{ object with the genomic coordinates
        of the sequences to be searched.}
}}

\item{...}{Additional arguments for specific methods.}

\item{method}{The internal method to use for motif searching. One of
\itemize{
    \item{\code{"homer2"}}{ call to the homer2 binary}
    \item{\code{"matchPWM"}}{ using Biostrings::matchPWM}
}
Please note that the two methods might give slightly different results (see details).}

\item{homerfile}{Path and file name of the \code{homer2} binary.}

\item{Ncpu}{Number of parallel threads (only for \code{method="homer2"}, default: 1).}
}
\value{
A \code{GRanges} object with the matches to \code{query} in \code{subject}.
}
\description{
\code{findMotifHits} scans sequences (either provided
as a file, an R object or genomic coordinates)  for matches to
positional weight matrices (provided as a file or as R objects)
}
\details{
The two implemented methods (\code{matchPWM} and \code{homer2}) are there
    for convenience (\code{matchPWM} is available as an R package and does not
    have any software dependencies outside of R) and for efficiency (\code{homer2}
    is typically faster but requires a separate installation of Homer).

    In general, running \code{findMotifHits} with the same parameters using the two
    methods generates identical results. Some minor differences could occur and result
    from rounding errors during the necessary conversion of PWMs (log2-odd scores) to
    the probability matrices needed by Homer, and the conversion of score from and to
    the natural log scale used by Homer. These conversions are implemented transparently
    for the user, so that the arguments of \code{findMotifHits} do not have to be
    adjusted (e.g. the PWMs shold always contain log2-odd scores, and \code{min.score}
    is always on the log2 scale).
}
\examples{
\dontrun{
findMotifHits(pwm, "sequences.fa")
}
}
\seealso{
\code{\link[Biostrings]{matchPWM}}, \code{\link[TFBSTools]{searchSeq}}
}
