% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{get_binned_motif_enrichment}
\alias{get_binned_motif_enrichment}
\title{Do Binned Motif Enrichment Analysis with `monaLisa` a la `Homer`}
\usage{
get_binned_motif_enrichment(
  seqs,
  bins,
  pwmL,
  genome = NULL,
  enrichment_test = c("binomial", "fishers_exact"),
  frac_N_allowed = 0.7,
  max_kmer_size = 3L,
  min.score = 10,
  match_method = "matchPWM.concat",
  Ncpu = 1L,
  verbose = TRUE
)
}
\arguments{
\item{seqs}{DNAStringSet object with sequences to test}

\item{bins}{factor vector of the same length and order as \code{seqs}, indicating the bin
each sequence belongs to. See the \code{\link[monaLisa]{bin}} function.}

\item{pwmL}{PWMatrixList object with PWMs of motifs.}

\item{genome}{a BSgenome object indicating the genome to which the \code{seqs} 
belong to, and which will be scanned for motif matches.}

\item{enrichment_test}{A \code{character} scalar specifying the type of
enrichment test to perform. One of \code{"binomial"} (default) or
\code{"fisher_exact"}. The enrichment test is one-sided (enriched in foreground).}

\item{frac_N_allowed}{A numeric scalar with the maximal fraction of N bases allowed 
in a sequence (defaults to 0.7). Sequences with a fraction of N bases greater than
this will be excluded from the analysis.}

\item{max_kmer_size}{the maximum kmer size to consider, when adjusting background weights
for kmer composition compared to the foreground sequences (default 1-mer, 2-mers and 3-mers).}

\item{min.score}{the \code{min.score} parameter for the \code{findMotifHits} function. It
is the minimum score for counting a match (default set to 10).}

\item{match_method}{the \code{method} parameter in \code{findMotifHits}. It specifies
the method used for motif searching.}

\item{Ncpu}{Number of CPUs to use (default set to 1). This can enter the \code{findMotifHits}
function when searching for motif matches.}

\item{verbose}{A logical scalar. If \code{TRUE}, report progress.}
}
\description{
This function does a motif enrichment analysis on the set of sequences belonging
  to the same bin, using all the sequences in the rest of the bins as background. 
  In each enrichment analysis (per bin), \code{get_binned_motif_enrichment} uses other 
  functions within `monaLisa`, which do a motif enrichment analysis as `Homer` (version 4.11)
  implements. See Details for more.
}
\details{
TODO: - Use Ncpu to also parallelize across list of DFs (enrichment per bin)
               - parallellize the other functions using Ncpu.
               - add log-enrichment info to final output.
               - for fisher's exact test: update dosumentation on one-sidedness.
               - change functions to do one-time calculations when using in binned mode?
               - filter seqs: make one-time thing.
               - should the calculated p-values be adjusted for multiple testing? (something Homer doesn't do)
               - avoid ordering of TFs in enrichment, to more easily patch things up together.
}
