% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{get_motif_enrichment}
\alias{get_motif_enrichment}
\title{Do Motif Enrichment}
\usage{
get_motif_enrichment(motif_matrix = NULL, df = NULL)
}
\arguments{
\item{motif_matrix}{matrix with 0 and 1 entries for absence or presence of 
motif hits per sequence.}

\item{df}{DatFrame object (output of running \code{iterate_norm_for_kmer_comp})}
}
\value{
a data.frame containing the motif names and the log10(p-value) of their enrichment
}
\description{
Do Motif Enrichment
}
