% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{get_motif_enrichment}
\alias{get_motif_enrichment}
\title{Do Motif Enrichment}
\usage{
get_motif_enrichment(
  motif_matrix = NULL,
  df = NULL,
  test = c("binomial", "fishers_exact"),
  verbose = TRUE
)
}
\arguments{
\item{motif_matrix}{matrix with 0 and 1 entries for absence or presence of 
motif hits per sequence.}

\item{df}{DatFrame object (output of running \code{iterate_norm_for_kmer_comp})}

\item{test}{type of test to do for the motif enrichment. By default it is the binomial, which 
is what \code{Homer} uses by default. Fisher's exact test is another 
option which allows for testing enrichment, without having to account for special cases of zero 
background counts for a motif. \code{fisher.test} is used with \code{alternative="greater"}, making
it a one-sided test for enrichment, as is the case with the binomial test.}

\item{verbose}{A logical scalar. If \code{TRUE}, report motif enrichment test.}
}
\value{
a \code{data.frame} containing the motifs as rows and the following columns: \itemize{
  \item{motif_name}{: the motif name}
  \item{log_p_value}{: the log p-value for enrichment. If test=binomial (default), this log p-value is identical to the one
     that Homer returns.}
  \item{fg_weight_sum}{: the sum of the weights of the foreGround sequences that have at least one instance of a specific motif (ZOOPS mode).}
  \item{bg_weight_sum}{: the sum of the weights of the backGround sequences that have at least one instance of a specific motif (ZOOPS mode).}
  \item{fg_weight_sum_total}{: the total sum of the weights of all foreGround sequences.}
  \item{bg_weight_sum_total}{: the total sum of the weights of all backGround sequences.}
}
}
\description{
Do Motif Enrichment
}
