% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getColsByBin}
\alias{getColsByBin}
\title{Get colors by bin.}
\usage{
getColsByBin(b, col1 = c("#3F007D", "#54278F", "#6A51A3", "#807DBA",
  "#9E9AC8", "#BCBDDC", "#DADAEB"), col2 = c("#FDD0A2", "#FDAE6B",
  "#FD8D3C", "#F16913", "#D94801", "#A63603", "#7F2704"),
  col0 = "AAAAAA33")
}
\arguments{
\item{b}{A factor that groups elements into bins (typically the output of
\code{\link{bin}}).}

\item{col1}{First color.}

\item{col2}{Second color.}

\item{col0}{Neutral color.}
}
\value{
A character vector with colors for the elements in \code{b}.
}
\description{
Get colors for elements according to their bin.

Colors are assigned to bins forming a gradient from \code{col1}
    to \code{col2} in the order of \code{levels{b}}. \code{col0} is assigned
    to the neutral bin (attribute \code{""}) if available.
}
\seealso{
\code{\link{bin}}.
}
