% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{getGCweight}
\alias{getGCweight}
\title{Get GC weights for background regions}
\usage{
getGCweight(inputList = NULL)
}
\arguments{
\item{inputList}{a `list` containing
\itemize{
  \item{sequenceWeights}{: a \code{dataframe} with nrows equal to the number of sequences, with a 
    column called `foreGround` that is 1 if a sequence is in the foreGround group, and 0 otherwise, 
    and a column containing the name of the sequnce.}
  \item{sequenceNucleotides}{: a \code{DNAStringSet} object containing the raw sequences}
}}
}
\value{
a list containing:
  \itemize{
    \item{sequenceWeights}{: a \code{dataframe} containing the sequence GC content, GC bins
      they were assigned to, and the weight to correct for GC differences between foreGround
      and backGround sequences}
    \item{sequenceNucleotides}{: a \code{DNAStringSet} object containing the raw sequences}
  }
}
\description{
In this implementation we still follow HOMER. But this can be re-implemented in 
  a more continuous fashion. Each sequence is put in a specific GC bin depending on its GC content. 
  Then, for each GC bin, the number of foreGround and backGround sequences in that bin is calculated.
  Weights are calculated for the backGround sequences in bin i as follows:
  weight_i = (number_fg_seqs_i/number_bg_seqs_i)*(number_bg_seqs_total/number_fg_seqs_total)
}
