% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{getKmerFreq}
\alias{getKmerFreq}
\title{Calculate observed and expected k-mer frequencies}
\usage{
getKmerFreq(seqs, kmerLen = 4, MMorder = 2, pseudoCount = 1)
}
\arguments{
\item{seqs}{Set of sequences, either a \code{character} vector or a
\code{\link{DNAStringSet}}.}

\item{kmerLen}{A \code{numeric} scalar giving the k-mer length.}

\item{MMorder}{A \code{numeric} scalar giving the order of the Markov model
used to calculate the expected frequencies.}

\item{pseudoCount}{A \code{numeric} scalar - will be added to the observed
counts for each k-mer to avoid zero values.}
}
\description{
Given a set of sequences, calculate observed and expected k-mer
  frequencies. Expected frequencies are based on a Markov model of order
  \code{MMorder}.
}
