% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{getMotifsFromDirGraph}
\alias{getMotifsFromDirGraph}
\title{Infer putative motifs from directed graph}
\usage{
getMotifsFromDirGraph(seqs, g, BPPARAM = SerialParam())
}
\arguments{
\item{seqs}{A \code{DNAStringSet} with sequences}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}

\item{A}{directed k-mer graph, e.g. from \code{buildDirGraphFromKmers}
or \code{filterDirGraph}}
}
\description{
Infer putative motifs from a directed graph
}
