% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_kmers.R
\name{getMotifsFromDirGraph}
\alias{getMotifsFromDirGraph}
\title{Infer putative motifs from directed graph}
\usage{
getMotifsFromDirGraph(seqs, g, BPPARAM = SerialParam(), includeRevComp = TRUE)
}
\arguments{
\item{seqs}{A \code{DNAStringSet} with sequences}

\item{g}{A directed k-mer graph, e.g. from \code{buildDirGraphFromKmers}
or \code{filterDirGraph}}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}

\item{includeRevComp}{A \code{logical} scalar. If \code{TRUE} (default),
concatenate \code{seqs} and their reverse-complemented versions
before building the graph. The choice here should correspond to the choice
made when extracting the k-mers.}
}
\description{
Infer putative motifs from a directed graph
}
