% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{is_valid_df}
\alias{is_valid_df}
\title{Check input object}
\usage{
is_valid_df(x)
}
\arguments{
\item{x}{Input object to be checked.}
}
\value{
\code{TRUE} if \code{x} is valid, \code{FALSE} otherwise
}
\description{
Check if the input object is valid, i.e. is of the correct
  object type (DataFrame) and has all expected columns and attributes.
}
