% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{norm_for_kmer_comp}
\alias{norm_for_kmer_comp}
\title{Adjust for k-mer composition (single iteration)}
\usage{
norm_for_kmer_comp(df, max_kmer_size = 3L, minimum_seq_weight = 0.001)
}
\arguments{
\item{df}{a \code{DataFrame} as returned by \code{calculate_GC_weight}.}

\item{max_kmer_size}{Integer scalar giving the maximum k-mer size to
consider. The default is set to 3 (like in \code{HOMER}), meaning that
k-mers of size 1, 2 and 3 are considered.}

\item{minimum_seq_weight}{Numeric scalar greater than zero giving the
minimal weight of a sequence. The default value (0.001) was also used by
\code{HOMER} (HOMER_MINIMUM_SEQ_WEIGHT  constant in Motif2.h).}
}
\value{
a \code{DataFrame} of the same dimensions as the input \code{df},
  with the weight to adjust for k-mer composition and the current error
  stored in the column \code{kmer_weight} and the attribute \code{err}.
}
\description{
Here we correct the background sequence weights, adjusting for
  k-mer composition compared to the foreground sequences. This function
  implements a single iteration, and is called iteratively by
  \code{iterate_norm_for_kmer_comp} to get to the final set of adjusted
  weights, which will be the result of adjusting for GC and k-mer
  composition. We closely follow HOMER's \code{normalizeSequenceIteration()}
  function found in \code{Motif2.cpp}.
}
