% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_HOMER.R
\name{parseHomerOutput}
\alias{parseHomerOutput}
\title{load output from HOMER findMotifsGenome.pl into R}
\usage{
parseHomerOutput(infiles, pseudocount.log2enr = 8, p.adjust.method = "BH")
}
\arguments{
\item{infiles}{HOMER output files to be parsed.}

\item{pseudocount.log2enr}{A numerical scalar with the pseudocount to add to
foreground and background counts when calculating log2 motif enrichments}

\item{p.adjust.method}{A character scalar selecting the p value adjustment
method (used in \code{\link[stats]{p.adjust}}).}
}
\value{
A list of four components (\code{negLog10P}, \code{negLog10Padj},
    \code{pearsonResid} and \code{log2enr}), containing each a motif (rows)
    by bin (columns) matrix with raw -log10 P values, -log10 adjusted P values,
    and motif enrichments as Pearson residuals (\code{pearsonResid}) and as
    log2 ratios (\code{log2enr}).
}
\description{
Parse HOMER output files into R data structures.
}
