% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_HOMER.R
\name{parseHomerOutput}
\alias{parseHomerOutput}
\title{load output from HOMER findMotifsGenome.pl into R}
\usage{
parseHomerOutput(
  infiles,
  pseudocount.log2enr = 8,
  pseudofreq.pearsonResid = 0.001,
  p.adjust.method = "BH"
)
}
\arguments{
\item{infiles}{HOMER output files to be parsed.}

\item{pseudocount.log2enr}{A numerical scalar with the pseudocount to add to
foreground and background counts when calculating log2 motif enrichments}

\item{pseudofreq.pearsonResid}{A numerical scalar with the pseudo-frequency
to add to background frequencies when calculating Pearson residuals.
The value needs to be in [0,1] and corresponds to the minimal expected
fraction of background sequences that contain at least one motif hit.}

\item{p.adjust.method}{A character scalar selecting the p value adjustment
method (used in \code{\link[stats]{p.adjust}}).}
}
\value{
A list of four components (\code{negLog10P}, \code{negLog10Padj},
    \code{pearsonResid} and \code{log2enr}), containing each a motif (rows)
    by bin (columns) matrix with raw -log10 P values, -log10 adjusted P values,
    and motif enrichments as Pearson residuals (\code{pearsonResid}) and as
    log2 ratios (\code{log2enr}).
}
\description{
Parse HOMER output files into R data structures.
}
\examples{
outfile <- system.file("extdata", "homer_output.txt.gz", package = "monaLisa")
res <- parseHomerOutput(infiles = c(bin1 = outfile))
head(res$negLog10P)

}
