% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBinDensity}
\alias{plotBinDensity}
\title{Density plot of binned elements.}
\usage{
plotBinDensity(
  x,
  b,
  xlab = deparse(substitute(x, env = as.environment(-1))),
  ylab = "Density",
  main = "",
  legend = "topright",
  legend.cex = 1,
  ...
)
}
\arguments{
\item{x}{A numerical vector with the values used for binning.}

\item{b}{A factor that groups elements of \code{x} into bins (typically the 
output of \code{\link{bin}}).}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{main}{Main title.}

\item{legend}{If not \code{NULL}, draw a legend with binning information
(will be passed to \code{legend(x=legend)} to control legend position).}

\item{legend.cex}{A scalar that controls the text size in the legend relative
to the current \code{par("cex")} (see \code{\link{legend}}).}

\item{...}{Further arguments passed to \code{\link{getColsByBin}}.}
}
\value{
Invisibly the return value of \code{density(x)} that generated the 
    plot.
}
\description{
Plot the density of binned elements with binning information.
}
\examples{
set.seed(1)
x <- rnorm(100)
b <- bin(x, "equalN", nElements = 10)
plotBinDensity(x, b)

}
\seealso{
\code{\link{getColsByBin}}
}
