% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBinHist}
\alias{plotBinHist}
\title{Histogram of binned elements.}
\usage{
plotBinHist(x, b, breaks = 10 * nlevels(b), xlab = deparse(substitute(x)),
  ylab = "Frequency", main = "", legend = "topright", ...)
}
\arguments{
\item{x}{A numerical vector with the values used for binning.}

\item{b}{A factor that groups elements of \code{x} into bins (typically the output of
\code{\link{bin}}).}

\item{breaks}{Controls the histogram breaks (passed to \code{hist(...)}).}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{main}{Main title.}

\item{legend}{If not \code{NULL}, draw a legend with binning information (will
be passed to \code{legend(x=legend)} to control legend position).}

\item{...}{Further arguments passed to \code{\link{getColsByBin}}.}
}
\value{
Invisibly the return value of \code{hist(...)} that generated the plot.
}
\description{
Plot a histogram of binned elements with binning information.
}
\seealso{
\code{\link{getColsByBin}}, \code{\link[graphics]{hist}}
}
