% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMotifDirectionality}
\alias{plotMotifDirectionality}
\title{Plot Directionality of Predictor Effect}
\usage{
plotMotifDirectionality(
  stabs_obj = NULL,
  response = NULL,
  predictor_matrix = NULL,
  sel_color = "cadetblue",
  min_sel_prob = 0.4,
  cor_method = "pearson",
  ...
)
}
\arguments{
\item{stabs_obj}{the \code{stabs} object resulting from stability selection.}

\item{response}{the response vector that was used for the stability selection (like the log-fold change of a measure of interest).}

\item{predictor_matrix}{the predictor matrix that was used for the stability selection (like the number of predicted TFBS of all motifs across the regions of interest).}

\item{sel_color}{the color for the selected predictors from stability selection.}

\item{min_sel_prob}{predictors with a selection probability greater than or equal to this are included in the plot.}

\item{cor_method}{the correlation method to be used.}

\item{...}{additional parameters for the \code{barplot} function.}
}
\value{
a barplot indicating the directionality of the motifs with respect to the correlation to the response vector.
}
\description{
This function plots the selectiong probabilities of the chosen predictors (for example the selected motifs)
and assigns a + or - sign to these probabilities to give a sense of directionality of the effect. The assumption is that 
the response vector on which stability selection was performed is a measure of fold-change. The correlation (pearson by default)
of each predictor to the response vector is calculated. The selection probabilities of the chosen predictors multiplied by the 
sign of the correlation is plotted to indicate the directionality.
}
\seealso{
\code{\link[graphics]{barplot}}
}
