% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMotifDirectionality}
\alias{plotMotifDirectionality}
\title{Plot Directionality of Predictor Effects}
\usage{
plotMotifDirectionality(
  se,
  selProbMin = metadata(se)$stabsel.params.cutoff,
  selProbMinPlot = 0.4,
  col = "cadetblue",
  method = c("pearson", "kendall", "spearman"),
  ...
)
}
\arguments{
\item{se}{The \code{SummarizedExperiment} object with the results from
stability selection (typically a call to \code{\link{randomized_stabsel}}).}

\item{selProbMin}{A numerical scalar in [0,1]. Predictors with a selection
probability greater than \code{selProbMin} are shown as colored bars. The
color is defined by the \code{col} argument.}

\item{selProbMinPlot}{A numerical scalar in [0,1] less than \code{selProbMin}.
Predictors with a selection probability greater than \code{selProbMinPlot}
are shown as gray bars, in addition to the colored ones with probabilities
greater than \code{selProbMin}.}

\item{col}{the color for the selected predictors from stability selection.}

\item{method}{A character scalar with the correlation method to use.
One of "pearson", "kendall" or "spearman" (see \code{\link[stats]{cor}}).}

\item{...}{additional parameters for the \code{barplot} function.}
}
\value{
a barplot indicating the directionality of the predictors (motifs) with respect to the correlation to the response vector.
}
\description{
This function plots the selection probabilities of predictors
  (for example the selected motifs) multiplied with either +1 or -1 to give a
  sense of both the strength and the directionality of the associated effects.
  The directionality is estimated from the sign of the correlation coefficient
  between each predictor and the response vector.
}
\seealso{
\code{\link[graphics]{barplot}}
}
