% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMotifHeatmaps}
\alias{plotMotifHeatmaps}
\title{Heatmap of motif enrichments.}
\usage{
plotMotifHeatmaps(
  x,
  which.plots = c("p", "enr", "FDR", "log2enr"),
  width = 4,
  col.enr = c("#053061", "#2166AC", "#4393C3", "#92C5DE", "#D1E5F0", "#F7F7F7",
    "#FDDBC7", "#F4A582", "#D6604D", "#B2182B", "#67001F"),
  col.sig = c("#FFF5EB", "#FEE6CE", "#FDD0A2", "#FDAE6B", "#FD8D3C", "#F16913",
    "#D94801", "#A63603", "#7F2704"),
  maxEnr = NULL,
  maxSig = NULL,
  highlight = NULL,
  cluster = FALSE,
  show_dendrogram = FALSE,
  show_motif_GC = FALSE,
  show_seqlogo = FALSE,
  width.seqlogo = 1.5
)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} with numerical matrices
(motifs-by-bins) in its \code{assays()}, typically the return value
of \code{\link{runHomer}}.}

\item{which.plots}{Selects which heatmaps to plot (one or several from \code{"p"}, \code{"FDR"},
\code{"enr"} and \code{"log2enr"}).}

\item{width}{The width (in inches) of each individual heatmap, without legend.}

\item{col.enr}{Colors used for enrichment heatmap.}

\item{col.sig}{Colors used for significance hetmaps (P values and FDR).}

\item{maxEnr}{Cap color mapping at enrichment = \code{maxEnr} (default: 99.5th percentile).}

\item{maxSig}{Cap color mapping at -log10 P value or -log10 FDR = \code{maxSig}
(default: 99.5th percentile).}

\item{highlight}{A logical vector indicating motifs to be highlighted.}

\item{cluster}{If \code{TRUE}, the order of transcription factors will be determined by
hierarchical clustering of the \code{"enr"} component. Alternatively, an
\code{hclust}-object can be supplied which will determine the motif ordering.
No reordering is done for \code{cluster = FALSE}.}

\item{show_dendrogram}{If \code{cluster != FALSE}, controls whether to show
a row dendrogram for the clustering of motifs. Ignored for \code{cluster = FALSE}.}

\item{show_motif_GC}{If \code{TRUE}, show a column with the percent G+C of the motif
as part of the heatmap.}

\item{show_seqlogo}{If \code{TRUE}, show a sequence logo next to each motif label.
This will likely only make sense for a heatmap with a low number of motifs.}

\item{width.seqlogo}{The width (in inches) for the longest sequence logo (shorter
logos are drawn to scale).}
}
\value{
A list of \code{ComplexHeatmap::Heatmap} objects.
}
\description{
Plot motif enrichments (e.g. significance or magnitude) as a heatmap.
}
\details{
The heatmaps are plotted side-by-side and are created internally using
    the \pkg{ComplexHeatmap} package.

    Each heatmap will be \code{width} inches wide, so the total plot needs a
    graphics device with a width of at least \code{length(which.plots) * width}
    plus the space used for motif names and legend. The height will be auto-adjusted to
    the graphics device.
}
\references{
Gu, Z. Complex heatmaps reveal patterns and correlations in multidimensional
    genomic data. Bioinformatics 2016.
}
\seealso{
\code{\link{bin}}, \code{\link[ComplexHeatmap]{Heatmap}}
}
