% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSelectionProb}
\alias{plotSelectionProb}
\title{Plot selection probabilities of predictors}
\usage{
plotSelectionProb(
  se,
  directional = TRUE,
  selProbMin = metadata(se)$stabsel.params.cutoff,
  selProbMinPlot = 0.4,
  showSelProbMin = TRUE,
  selColor = "cadetblue",
  notSelColor = "grey",
  method = c("pearson", "kendall", "spearman"),
  ylimext = 0.2
)
}
\arguments{
\item{se}{The \code{SummarizedExperiment} object with the results from
stability selection (typically returned by \code{\link{randLassoStabSel}}).}

\item{directional}{A logical scalar. If \code{TRUE}, selection probabilities
are plotted with the sign of the marginal correlation between a predictor
and the response.}

\item{selProbMin}{A numerical scalar in [0,1]. Predictors with a selection
probability greater than \code{selProbMin} are considered selected and 
colored by the input from \code{selColor}. By default, \code{selProbMin} is
extracted from the parameters stored in \code{se}.}

\item{selProbMinPlot}{A numerical scalar in [0,1] less than 
\code{selProbMin}. Predictors with a selection probability greater 
than \code{selProbMinPlot} but less than \code{selProbMin} are shown as 
bars with color defined in \code{notSelColor}. \code{selProbMinPlot} is 
useful in order to include additional predictors in the barplot, that were 
not selected according to \code{selProbMin} but may be close to that cutoff
or are simply nice to contrast with the selected predictors. Setting 
\code{selProbMinPlot = 0} will of course include all predictors.}

\item{showSelProbMin}{A logical scalar. If \code{TRUE}, the value of
\code{selProbMin} is shown by a horizontal red line.}

\item{selColor}{Color for the selected predictors which have a selection
probability greater than \code{selProbMin}.}

\item{notSelColor}{Color for the rest of the (unselected) predictors which 
   will be show in the barplot.
A color vector giving the three colors used for predictors with
  selection probability greater than \code{selProbMin}, additional predictors
  with selection probability greater than \code{selProbMinPlot}, and the
  selection probability cutoff line.}

\item{method}{A character scalar with the correlation method to use in the
calculation of predictor-response marginal correlations. One of "pearson",
"kendall" or "spearman" (see \code{\link[stats]{cor}}).}

\item{ylimext}{A numeric scalar defining how much the y axis limits should be
expanded beyond the plotted probabilities to allow for space for the
bar labels. This value can be increased if the predictor names above the 
bars are too big and not showing in the plot.}
}
\value{
a \code{ggplot2} object.
}
\description{
This function plots the selection probabilities of predictors
  (for example the selected motifs), optionally multiplied with either +1 or
  -1 to give a sense of both the strength and the directionality of the
  associated effects. The directionality is estimated from the sign of the
  correlation coefficient between each predictor and the response vector.
}
\details{
This function creates a bar plot with \code{ggplot}.
  Each bar corresponds to a predictor (motif) and the colors correspond to 
  whether or not it was selected. The y-axis shows the selection 
  probabilities (\code{directional=FALSE}) or selection probabilities with 
  the sign of the marginal correlation to the response 
  (\code{directional=TRUE}).
}
\examples{
## create data set
Y <- rnorm(n = 500, mean = 2, sd = 1)
X <- matrix(data = NA, nrow = length(Y), ncol = 50)
for (i in seq_len(ncol(X))) {
  X[ ,i] <- runif(n = 500, min = 0, max = 3)
}
s_cols <- sample(x = seq_len(ncol(X)), size = 10, 
  replace = FALSE)
for (i in seq_along(s_cols)) {
  X[ ,s_cols[i]] <- X[ ,s_cols[i]] + Y
}
  
## reproducible randLassoStabSel() with 1 core
set.seed(123)
ss <- randLassoStabSel(x = X, y = Y)
plotSelectionProb(ss)

}
