% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSelectionProb}
\alias{plotSelectionProb}
\title{Barplot Selection Probabilities}
\usage{
plotSelectionProb(
  se,
  cutoff = metadata(se)$stabselParams$cutoff,
  ylim = c(0, 1.1),
  onlySelected = TRUE,
  sel_color = "cadetblue",
  las = 2,
  ...
)
}
\arguments{
\item{se}{the \code{SummarizedExperiment} object resulting from stability selection.}

\item{cutoff}{the selection probability cutoff.}

\item{ylim}{the limits for the y-axis.}

\item{onlySelected}{logical (default=TRUE) indicating if only selected predictors' selection probabilities
should be plotted.}

\item{sel_color}{color of the selected predictors.}

\item{las}{(2 by default) plot labels vertically or horizontally.}

\item{...}{additional parameters for the \code{barplot} function.}
}
\value{
barplot of selection probabilities.
}
\description{
Create a bar plot of the selection probabilities in descending order.
}
\seealso{
\code{\link[graphics]{barplot}}
}
