% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSelectionProb}
\alias{plotSelectionProb}
\title{Barplot Selection Probabilities}
\usage{
plotSelectionProb(
  se,
  selProbMin = metadata(se)$stabsel.params.cutoff,
  ylim = c(0, 1.1),
  onlySelected = TRUE,
  col = "cadetblue",
  las = 2,
  ...
)
}
\arguments{
\item{se}{the \code{SummarizedExperiment} object resulting from stability selection 
(typically a call to \code{\link{randLassoStabSel}}).}

\item{selProbMin}{A numerical scalar in [0,1]. Predictors with a selection
probability greater than \code{selProbMin} are shown as colored bars. The
color is defined by the \code{col} argument.}

\item{ylim}{the limits for the y-axis.}

\item{onlySelected}{logical (default=TRUE) indicating if only selected predictors' selection probabilities
should be plotted.}

\item{col}{color of the selected predictors.}

\item{las}{(2 by default) plot labels vertically or horizontally.}

\item{...}{additional parameters for the \code{barplot} function.}
}
\value{
barplot of selection probabilities.
}
\description{
Create a bar plot of the selection probabilities in descending order.
}
\seealso{
\code{\link[graphics]{barplot}}
}
