% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotStabilityPaths}
\alias{plotStabilityPaths}
\title{Plot Stability Paths}
\usage{
plotStabilityPaths(
  se,
  selProbMin = metadata(se)$stabsel.params.cutoff,
  col = "cadetblue",
  lwd = 1,
  lty = 1,
  ylim = c(0, 1.1),
  ...
)
}
\arguments{
\item{se}{the \code{SummarizedExperiment} object resulting from stability selection, 
by running \code{\link[monaLisa]{randomized_stabsel}}.}

\item{selProbMin}{A numerical scalar in [0,1]. Predictors with a selection
probability greater than \code{selProbMin} are shown as colored lines. The
color is defined by the \code{col} argument.}

\item{col}{color of the selected predictors.}

\item{lwd}{line width (default = 1).}

\item{lty}{line type (default = 1).}

\item{ylim}{limits for y-axis (default = c(0,1.1)).}

\item{...}{additional parameters to pass on to \code{matplot}.}
}
\value{
plot of stability paths.
}
\description{
Plot the stability paths of each variable (predictor), showing the selection probability
  as a function of the regularization step.
}
\seealso{
\code{\link[stabs]{stabsel}} and \code{\link[graphics]{matplot}}
}
