% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotStabilityPaths}
\alias{plotStabilityPaths}
\title{Plot Stability Paths}
\usage{
plotStabilityPaths(stabs_object, cols = NULL, lwd = 1, lty = 1,
  ylim = c(0, 1.1), ...)
}
\arguments{
\item{stabs_object}{the \code{stabs} object resulting from stability selection.}

\item{cols}{color vector for the varaiables from the predictor matrix. By default, it's set to NULL
and the function colors the variables by whether or not they were selected.}

\item{lwd}{line width (default 1).}

\item{lty}{line type (default 1).}

\item{ylim}{limits for y-axis (default c(0,1.1)).}

\item{...}{additional parameters to pass on to \code{matplot}.}
}
\value{
plot of stability paths.
}
\description{
Plot the stability paths of each variable (predictor), showing the selection probability
as a function of the regularization step.
}
\seealso{
\code{\link[stabs]{stabsel}} and \code{\link[graphics]{matplot}}
}
