% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_selection.R
\name{randomized_stabsel}
\alias{randomized_stabsel}
\title{Randomized Lasso Stability Selection}
\usage{
randomized_stabsel(x = x, y = y, weakness = 0.8, cutoff = 0.8,
  PFER = 2, ...)
}
\arguments{
\item{x}{the predictor matrix}

\item{y}{the response vector}

\item{weakness}{the parameter essential for the randomized lasso.}

\item{cutoff}{value between 0 and 1 which is the cutoff for the selection probability. Any variable
with a selection probability that is higher than the set cutoff will be selected.}

\item{PFER}{the absolute number of false positives that we allow for in the final list of 
selected variables. This comes from the Meinshausen and Buehlmann paper.}
}
\value{
a \code{stabsel} object
}
\description{
This function runs randomized lasso stability selection as presented by Meinshausen
and Bühlmann (2010) and with the improved error bounds introduced by Shah and Samworth (2013). The function 
uses the \code{stabsel} function from the \code{stabs} package, but using the randomized lasso version.
}
\details{
Randomized lasso stability selection runs randomized lasso several times on several subsamples of 
the response variable and predictor matrix (stability selection). N/2 elements from the response variable are
randomly chosen, where N is the length of the vector. Their corresponsing section of the predictor matrix is 
also chosen, and the \code{glmnet.randomized_lasso} function is applied. This is done multiple times, and 
results in selection probabilities for each predictor. The probability of a specific predictor is the number of 
times it was selected divided by the total number of subsamples that were made (total number of times the 
regression was performed). 

We make use of the \code{stabs} package that implements lasso stability selection, and adapt it to run the 
randomized lasso stability selection. The output is an object of type \code{stabsel}.
}
\author{
Dania Machlab
}
