% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment_monaLisa.R
\name{run_monaLisa}
\alias{run_monaLisa}
\title{Do Motif Enrichment Analysis with `monaLisa` (for simple fg vs bg
  situation, this will be changed for bins later)}
\usage{
run_monaLisa(
  seqs,
  is_foreground,
  pwmL,
  enrichment_test = c("binomial", "fishers_exact"),
  Ncpu = 1L,
  verbose = TRUE
)
}
\arguments{
\item{seqs}{DNAStringSet object with sequences to analyze}

\item{pwmL}{PWMatrixList object with PWMs of motifs}

\item{Ncpu}{Number of CPUs to use (default set to 1).}

\item{verbose}{A logical scalar. If \code{TRUE}, report on k-mer composition
adjustment.}

\item{is_forground}{logical vector of the same length as \code{seqs}.
\code{TRUE} indicates that the sequence corresponds to the foreground set,
\code{FALSE} indicates a background set sequence.}
}
\description{
Do Motif Enrichment Analysis with `monaLisa` (for simple fg vs bg
  situation, this will be changed for bins later)
}
\details{
TODO: - parallellize the other functions using Ncpu.
               - add log-enrichment info to final output.
               - for fisher's exact test: discriminate between enriched and depleted motifs
               - change functions to do one-time calculations when using in binned mode.
               - make a better get_motif_hits_in_ZOOPS_mode function with our own 
                 implementation of matchPWM (faster). for now it uses homer2 because it's so fast.
               - should the calculated p-values be adjusted for multiple testing? (something Homer doesn't do)
}
