% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_random_regions.R
\name{sample_random_regions}
\alias{sample_random_regions}
\title{Sample random regions.}
\usage{
sample_random_regions(
  N = 100L,
  regWidth = 200L,
  fractionCGI = 0.5,
  mappableRegions = NULL,
  seed = 123
)
}
\arguments{
\item{N}{Number of regions to sample.}

\item{regWidth}{Region width.}

\item{fractionCGI}{fraction of regions to be sampled from CpG islands. Not implemented yet.}

\item{mappableRegions}{A unstranded GRanges object of the mappable parts of the genome. This can be a GRanges of
of the entire genome with one range from start to end of each chromosome if mappability
is not an issue.}

\item{seed}{Seed for seeding the random number generator.}
}
\value{
A GRanges object with randomly sampled mappable regions of width \code{regWidth}
        with \code{fractionCGI} coming from CpG islands.
}
\description{
Sample random regions from the mappable parts of the genome with
     a given fraction from CpG islands.
}
\details{
to be filled in.
}
