% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_random_regions.R
\name{sample_random_regions}
\alias{sample_random_regions}
\title{Sample random regions of fixed length.}
\usage{
sample_random_regions(
  allowedRegions = NULL,
  N = 100L,
  regWidth = 200L,
  seed = 123
)
}
\arguments{
\item{allowedRegions}{An unstranded GRanges object of the "allowed" of the genome, usually the mappable regions.}

\item{N}{Number of regions to sample.}

\item{regWidth}{Region width.}

\item{seed}{Seed for seeding the random number generator.}
}
\value{
A GRanges object with randomly sampled mappable regions of width \code{regWidth}
        with \code{fractionCGI} coming from CpG islands.
}
\description{
Sample random regions from the mappable parts of the genome with
     a given fraction from CpG islands.
}
\details{
to be filled in.
}
