% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqlogo.R
\name{seqLogoGrob}
\alias{seqLogoGrob}
\title{Create a simple sequence logo grob.}
\usage{
seqLogoGrob(x)
}
\arguments{
\item{x}{A \code{\link[TFBSTools]{PFMatrix}} object}
}
\value{
A polygon grob.
}
\description{
Create a simple sequence logo grob (grid-graphics object) for a
    transcription factor from a position frequency matrix. The logo drawing
    code is a simplified version from \code{\link[seqLogo]{seqLogo}} and for
    example can be used to embedd sequence logos within other plots.
}
\seealso{
\code{\link[seqLogo]{seqLogo}} for the original, more flexible version
    of this function.
}
